+++ app/contact/success/page.tsx
@@ -1,5 +1,6 @@
-import SendLottie from "@/components/Lottie/SendLottie";
 import { buttonVariants } from "@/components/ui/button";
+import Link from "next/link";
+import SendLottieClient from "@/components/Lottie/SendLottieClient";
 
 export const metadata = {
   title: "전송 완료",
@@ -9,10 +10,10 @@ const SuccessPage = () => {
   return (
     <main className="mx-auto flex h-full max-w-sm flex-1 flex-col items-center justify-center gap-6 px-6">
       <h2 className="text-xl font-semibold">성공적으로 전송하였습니다.</h2>
-      <SendLottie />
-      <a href="/" className={buttonVariants({ variant: "default" })}>
+      <SendLottieClient />
+      <Link replace href="/" className={buttonVariants({ variant: "default" })}>
         홈으로 되돌아가기
-      </a>
+      </Link>
     </main>
   );
 };
