+++ components/Lottie/SendLottieClient.tsx
@@ -1,0 +1,11 @@
+"use client";
+import dynamic from "next/dynamic";
+
+const SendLottie = dynamic(() => import("@/components/Lottie/SendLottie"), {
+  ssr: false,
+});
+const SendLottieClient = () => {
+  return <SendLottie />;
+};
+
+export default SendLottieClient;
